pyproject.toml
[project]
name = "agentic-resume-intelligence-analyzer-aria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fpdf2>=2.8.5",
    "jq>=1.10.0",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langgraph>=1.0.2",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "prefect>=3.5.0",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.4",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.3",
    "streamlit>=1.51.0",
    "streamlit-audiorec>=0.1.3",
    "textblob>=0.19.0",
    "tiktoken>=0.12.0",
]
